GO: experiment opt layouut
