+++ libs/gi/localization/assets/locales/en/ui.json
@@ -5,6 +5,7 @@
     "weapons": "Weapons",
     "artifacts": "Artifacts",
     "characters": "Characters",
+    "optimize": "Experiment",
     "teams": "Teams",
     "tools": "Tools",
     "setting": "Settings",
