+++ libs/gi/page-optimize/README.md
@@ -1,0 +1,3 @@
+# gi-page-optimize
+
+Optimize-first experiment route for Genshin Optimizer (`#/experiment`).
