+++ libs/gi/page-team/src/CharacterDisplay/OptimizeChrome/OptimizeCalcBarContext.tsx
@@ -1,0 +1,19 @@
+import { createContext, useContext } from 'react'
+
+const OptimizeCalcBarContext = createContext(false)
+
+export function OptimizeCalcBarProvider({
+  children,
+}: {
+  children: React.ReactNode
+}) {
+  return (
+    <OptimizeCalcBarContext.Provider value={true}>
+      {children}
+    </OptimizeCalcBarContext.Provider>
+  )
+}
+
+export function useOptimizeCalcBar() {
+  return useContext(OptimizeCalcBarContext)
+}
