+++ libs/gi/page-team/src/loadoutTabPath.ts
@@ -1,0 +1,6 @@
+export {
+  getTeamCharTabPath,
+  isTeamCharTabSegment,
+  TEAM_CHAR_TAB_SEGMENTS,
+  type TeamCharTabSegment,
+} from '@genshin-optimizer/gi/ui'
