+++ libs/gi/ui/src/index.ts
@@ -6,3 +6,4 @@ export * from './context'
 export * from './hooks'
 export type * from './type'
 export * from './util'
+export * from './optimize'
