+++ libs/gi/ui/src/optimize/index.ts
@@ -1,0 +1,3 @@
+export * from './charTabPaths'
+export * from './ensureOptimizeContext'
+export * from './openOptimizeFlow'
